Removing all the old GDC-GRIN2 related files
