+++ rust/Cargo.toml
@@ -104,10 +104,6 @@ name="validateHDF5"
 path="src/validateHDF5.rs"
 
 [[bin]]
-name="gdcGRIN2"
-path="src/gdcGRIN2.rs"
-
-[[bin]]
 name="cerno"
 path="src/cerno.rs"
