--- server/src/grin2/test/gdc.grin2.unit.spec.ts
@@ -1,218 +1,0 @@
-import tape from 'tape'
-import fs from 'fs'
-import path from 'path'
-import serverconfig from '../../serverconfig.js'
-import zlib from 'zlib'
-import { parseGdcCnvFile } from '../../gdc.cnvFile.ts'
-import { parseGdcMafFile, gunzipIfNeeded } from '../../gdc.mafFile.ts'
-import { filterAndConvertCnv, processGdcCaseFiles } from '../main.ts'
-import { dtcnv, dtloh, dtsnvindel } from '#shared'
-import type { GRIN2Request } from '#types'
-
-/* test sections
-
-parseGdcCnvFile: real GDC segmean fixture
-GDC segmean → general route lesions match the Rust gdcGRIN2 golden output
-parseGdcCnvFile: copyNumber + LOH (allele-specific format)
-processGdcCaseFiles: unified-run assembly reproduces the Rust golden output
-gunzipIfNeeded: real GDC MAF files are gzipped; plain text passes through
-parseGdcMafFile: real GDC MAF fixture
-processGdcCaseFiles: MAF path emits mutation lesions with consequence + depth/altcount filtering
-*/
-
-/*
-These specs exercise the GDC GRIN2 ingestion layer — the CNV and MAF files the GDC route fetches and
-(in the legacy path) fed to the Rust `gdcGRIN2` binary. The CNV specs use the same fixture
-(server/test/tp/files/Grin2Test/TCGA-CG-4300.cnv.seg.txt), proving the general route's CNV path
-(parseGdcCnvFile → filterAndConvertCnv) reproduces the Rust lesion output; the MAF specs cover gzip
-handling, parsing, and the consequence + depth/altcount filtering done in the unified run. The golden
-output (Test_GRIN2_output.json) was generated by rust/test/gdcGRIN2.unit.spec.js with
-cnvOptions { lossThreshold: -0.4, gainThreshold: 0.3, segLength: 2_000_000 }.
-*/
-
-const FIXDIR = path.join(serverconfig.tpmasterdir, 'files/Grin2Test')
-const CASE_ID = 'TCGA-CG-4300' // case submitter id used as the lesion sample id by the GDC route
-// match the thresholds the Rust golden output was generated with
-const cnvOpts = { lossThreshold: -0.4, gainThreshold: 0.3, maxSegLength: 2_000_000 }
-
-// stable key for order-insensitive lesion comparison
-const lesionKey = (l: any[]) => [l[1], Number(l[2]), Number(l[3]), l[4]].join('|')
-
-tape('\n', function (test) {
-	test.comment('-***- server/grin2 GDC ingestion specs (CNV + MAF) -***-')
-	test.end()
-})
-
-tape('parseGdcCnvFile: real GDC segmean fixture', test => {
-	const content = fs.readFileSync(path.join(FIXDIR, 'TCGA-CG-4300.cnv.seg.txt'), { encoding: 'utf8' })
-	const mlst = parseGdcCnvFile(content)
-	test.ok(mlst.length > 0, 'parses a non-empty mlst from the GDC segmean file')
-	test.ok(
-		mlst.every(m => m.dt === dtcnv && m.valueType === 'segmean' && Number.isFinite(m.value)),
-		'every entry is dtcnv with valueType "segmean" and a finite numeric value'
-	)
-	test.end()
-})
-
-tape('GDC segmean → general route lesions match the Rust gdcGRIN2 golden output', test => {
-	const content = fs.readFileSync(path.join(FIXDIR, 'TCGA-CG-4300.cnv.seg.txt'), { encoding: 'utf8' })
-	const mlst = parseGdcCnvFile(content)
-
-	// classify each parsed segment through the general route's CNV converter
-	const ours: any[] = []
-	for (const m of mlst) {
-		const les = filterAndConvertCnv(CASE_ID, m, cnvOpts, 'segmean')
-		if (les) ours.push(les)
-	}
-
-	// golden: the Rust lesion output; keep only cnv (gain/loss) lesions for this case
-	const golden = JSON.parse(fs.readFileSync(path.join(FIXDIR, 'Test_GRIN2_output.json'), { encoding: 'utf8' }))
-	const goldenLesions: any[][] = JSON.parse(golden.grin2lesion)
-	const goldenCnv = goldenLesions.filter(l => l[4] === 'gain' || l[4] === 'loss')
-
-	test.ok(
-		ours.length > 0 && goldenCnv.length > 0,
-		`non-empty lesion sets (ours ${ours.length}, golden ${goldenCnv.length})`
-	)
-	test.equal(ours.length, goldenCnv.length, 'same number of gain/loss lesions as Rust gdcGRIN2')
-
-	const oursSorted = ours.map(lesionKey).sort()
-	const goldenSorted = goldenCnv.map(lesionKey).sort()
-	test.deepEqual(oursSorted, goldenSorted, 'general route reproduces the exact Rust gain/loss lesions')
-
-	// spot-check a known classification from the fixture: chr1:2652001-2757000 segmean -1.344049 => loss
-	test.ok(
-		ours.some(l => l[1] === 'chr1' && Number(l[2]) === 2652001 && Number(l[3]) === 2757000 && l[4] === 'loss'),
-		'fixture segment with segmean -1.344049 classified as loss'
-	)
-	test.end()
-})
-
-tape('parseGdcCnvFile: copyNumber + LOH (allele-specific format)', test => {
-	// minimal allele-specific file: a balanced gain, and a one-allele-loss segment that should also emit LOH
-	const content = [
-		'GDC_Aliquot\tChromosome\tStart\tEnd\tCopy_Number\tMajor_Copy_Number\tMinor_Copy_Number',
-		'A1\tchr1\t1000\t2000\t4\t2\t2', // total CN 4, balanced → no LOH
-		'A1\tchr2\t5000\t6000\t1\t1\t0' // total CN 1, minor 0 & major>0 → LOH emitted
-	].join('\n')
-
-	const mlst = parseGdcCnvFile(content)
-	const cnvs = mlst.filter(m => m.dt === dtcnv)
-	const lohs = mlst.filter(m => m.dt === dtloh)
-
-	test.equal(cnvs.length, 2, 'two copyNumber segments parsed')
-	test.ok(
-		cnvs.every(m => m.valueType === 'copyNumber'),
-		'copyNumber segments tagged valueType "copyNumber"'
-	)
-	test.equal(cnvs[0].value, 4, 'value is the total copy number (4)')
-	test.equal(lohs.length, 1, 'one LOH entry emitted for the minor==0 & major>0 segment')
-	test.deepEqual(
-		[lohs[0].chr, lohs[0].start, lohs[0].stop],
-		['chr2', 5000, 6000],
-		'LOH entry carries the segment coordinates'
-	)
-
-	// the copyNumber segments classify correctly under copy-number thresholds (loss<=1, gain>=3, neutral=2)
-	test.deepEqual(
-		filterAndConvertCnv(CASE_ID, cnvs[0], { lossThreshold: 1, gainThreshold: 3, maxSegLength: 0 }, 'copyNumber'),
-		[CASE_ID, 'chr1', 1000, 2000, 'gain'],
-		'CN 4 >= gain threshold 3 => gain'
-	)
-	test.deepEqual(
-		filterAndConvertCnv(CASE_ID, cnvs[1], { lossThreshold: 1, gainThreshold: 3, maxSegLength: 0 }, 'copyNumber'),
-		[CASE_ID, 'chr2', 5000, 6000, 'loss'],
-		'CN 1 <= loss threshold 1 => loss'
-	)
-	test.end()
-})
-
-tape('processGdcCaseFiles: unified-run assembly reproduces the Rust golden output', async test => {
-	// drive the actual unified-run assembly with an injected loader so no GDC network is needed:
-	// the loader returns the parsed fixture mlst for the case's cnv file id.
-	const content = fs.readFileSync(path.join(FIXDIR, 'TCGA-CG-4300.cnv.seg.txt'), { encoding: 'utf8' })
-	const loadCnv = async (_cnvFileId: string) => parseGdcCnvFile(content)
-
-	const caseFiles = { [CASE_ID]: { cnv: 'fake-cnv-file-uuid' } }
-	const request = { caseFiles, cnvOptions: cnvOpts } as unknown as GRIN2Request
-
-	const { lesions, processing } = await processGdcCaseFiles(caseFiles, request, {} as any, undefined, { loadCnv })
-
-	// compare assembled lesions to the Rust golden cnv lesions
-	const golden = JSON.parse(fs.readFileSync(path.join(FIXDIR, 'Test_GRIN2_output.json'), { encoding: 'utf8' }))
-	const goldenCnv = (JSON.parse(golden.grin2lesion) as any[][]).filter(l => l[4] === 'gain' || l[4] === 'loss')
-
-	test.deepEqual(
-		lesions.map(lesionKey).sort(),
-		goldenCnv.map(lesionKey).sort(),
-		'unified-run assembly produces the exact Rust gain/loss lesions from caseFiles'
-	)
-	test.equal(processing.processedSamples, 1, 'one case processed')
-	test.equal(
-		processing.lesionCounts.byType.gain.count + processing.lesionCounts.byType.loss.count,
-		goldenCnv.length,
-		'processing summary gain+loss counts match the golden lesion count'
-	)
-	test.end()
-})
-
-tape('gunzipIfNeeded: real GDC MAF files are gzipped; plain text passes through', test => {
-	// Real GDC /data MAF bodies are .maf.gz (application/gzip, not Content-Encoding) — they must be
-	// gunzipped before parsing. The uncompressed test fixture masked this; this guards the live path.
-	const content = fs.readFileSync(path.join(FIXDIR, 'TCGA-CG-4300.maf.txt'), { encoding: 'utf8' })
-	const gz = zlib.gzipSync(Buffer.from(content))
-	test.ok(gz[0] === 0x1f && gz[1] === 0x8b, 'sanity: gzip buffer carries the gzip magic bytes')
-	test.equal(gunzipIfNeeded(gz), content, 'gzipped body is transparently decompressed to the original text')
-	test.equal(gunzipIfNeeded(Buffer.from(content)), content, 'plain-text body passes through unchanged')
-	test.equal(
-		parseGdcMafFile(gunzipIfNeeded(gz)).length,
-		parseGdcMafFile(content).length,
-		'parse count matches via gunzip path'
-	)
-	test.end()
-})
-
-tape('parseGdcMafFile: real GDC MAF fixture', test => {
-	const content = fs.readFileSync(path.join(FIXDIR, 'TCGA-CG-4300.maf.txt'), { encoding: 'utf8' })
-	const mlst = parseGdcMafFile(content)
-	test.ok(mlst.length > 0, 'parses a non-empty mlst from the GDC MAF file')
-	test.ok(
-		mlst.every(m => m.dt === dtsnvindel && Number.isInteger(m.pos) && typeof m.chr === 'string'),
-		'every entry is dtsnvindel with an integer pos and a chr'
-	)
-	test.ok(
-		mlst.every(m => Number.isFinite(m.totalDepth) && Number.isFinite(m.altCount)),
-		'each entry carries finite t_depth / t_alt_count'
-	)
-	test.ok(
-		mlst.some(m => m.class && m.class.length > 0),
-		'One_Consequence is captured as class for at least some rows'
-	)
-	test.end()
-})
-
-tape('processGdcCaseFiles: MAF path emits mutation lesions with consequence + depth/altcount filtering', async test => {
-	// inject a tiny MAF mlst so filtering is deterministic
-	const mafMlst = [
-		{ dt: dtsnvindel, chr: 'chr1', pos: 100, class: 'missense_variant', totalDepth: 30, altCount: 5 }, // kept
-		{ dt: dtsnvindel, chr: 'chr1', pos: 200, class: 'synonymous_variant', totalDepth: 30, altCount: 5 }, // wrong consequence
-		{ dt: dtsnvindel, chr: 'chr1', pos: 300, class: 'missense_variant', totalDepth: 5, altCount: 5 }, // depth < 10
-		{ dt: dtsnvindel, chr: 'chr1', pos: 400, class: 'missense_variant', totalDepth: 30, altCount: 1 } // alt < 2
-	]
-	const loadMaf = async (_id: string) => mafMlst
-	const caseFiles = { [CASE_ID]: { maf: 'fake-maf-uuid' } }
-	// consequences are pp mclass keys (what the shared GRIN2 UI emits); 'M' = missense. The filter
-	// translates each row's SO term (One_Consequence) to its mclass before comparing.
-	const request = {
-		caseFiles,
-		snvindelOptions: { consequences: ['M'], minTotalDepth: 10, minAltAlleleCount: 2 }
-	} as unknown as GRIN2Request
-
-	const { lesions } = await processGdcCaseFiles(caseFiles, request, {} as any, undefined, { loadMaf })
-	test.deepEqual(
-		lesions,
-		[[CASE_ID, 'chr1', 100, 100, 'mutation']],
-		'only the row passing consequence + depth + alt-count filters becomes a mutation lesion'
-	)
-	test.end()
-})
