+++ shared/types/src/index.ts
@@ -21,7 +21,6 @@ export * from './routes/errorResponse.ts'
 export * from './routes/filter.gdc.ts'
 export * from './routes/gdc.maf.ts'
 export * from './routes/gdc.mafBuild.ts'
-export * from './routes/gdc.grin2.ts'
 export * from './routes/grin2.ts'
 export * from './routes/genelookup.ts'
 export * from './routes/genesetEnrichment.ts'
